Remove a described carriage return
In this commit, I have removed one of the carriage returns in the PUT code snippet given. This is the case because the code snippet in before the change suggests that two carriage returns are below the `Content-length:12` line and before `Hello World!`. However,  instead, only one carriage return should be between the two, hence the removal.