Update Lecture-3.md
Fixed 2 typos. 

I would suggest that when talking about the basic constructor, instead of saying "objects do not share functions" one should say "objects do not share functions that are added on the fly". It seems more specific that way, because someone could argue that if we put a function inside a constructor, all the objects would share that function.